Update single device classes
