+++ src/waterHeater.test.ts
@@ -5,20 +5,21 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
 import { jest } from '@jest/globals';
-import { DeviceTypeId, VendorId, ServerNode, LogFormat as MatterLogFormat, LogLevel as MatterLogLevel, Environment } from '@matter/main';
-import { RootEndpoint } from '@matter/main/endpoints';
-import { MdnsService } from '@matter/main/protocol';
 import { AnsiLogger, LogLevel, TimestampFormat } from 'node-ansi-logger';
 import { rmSync } from 'node:fs';
 import path from 'node:path';
 
 import { Matterbridge } from './matterbridge.js';
 import { MatterbridgeEndpoint } from './matterbridgeEndpoint.js';
-import { WaterHeater } from './waterHeater.js';
+import { MatterbridgeWaterHeaterManagementServer, MatterbridgeWaterHeaterModeServer, WaterHeater } from './waterHeater.js';
+import { MatterbridgeThermostatServer } from './matterbridgeBehaviors.js';
+import { invokeBehaviorCommand } from './matterbridgeEndpointHelpers.js';
+
+import { DeviceTypeId, VendorId, ServerNode, LogFormat as MatterLogFormat, LogLevel as MatterLogLevel, Environment } from '@matter/main';
+import { RootEndpoint } from '@matter/main/endpoints';
+import { MdnsService } from '@matter/main/protocol';
 import { Identify, PowerSource, Thermostat, WaterHeaterManagement } from '@matter/main/clusters';
 import { ThermostatServer, WaterHeaterManagementServer, WaterHeaterModeServer } from '@matter/node/behaviors';
-import { MatterbridgeThermostatServer, MatterbridgeWaterHeaterManagementServer, MatterbridgeWaterHeaterModeServer } from './matterbridgeBehaviors.js';
-import { invokeBehaviorCommand } from './matterbridgeEndpointHelpers.js';
 
 let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
 let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
