+++ src/waterHeater.ts
@@ -25,10 +25,15 @@
 // Matterbridge
 import { MatterbridgeEndpoint } from './matterbridgeEndpoint.js';
 import { waterHeater } from './matterbridgeDeviceTypes.js';
-import { MatterbridgeWaterHeaterManagementServer, MatterbridgeWaterHeaterModeServer } from './matterbridgeBehaviors.js';
+import { MatterbridgeServer } from './matterbridgeBehaviors.js';
 
 // Matter.js
-import { WaterHeaterManagement, WaterHeaterMode } from '@matter/main/clusters';
+import { MaybePromise } from '@matter/main';
+import { ModeBase } from '@matter/main/clusters/mode-base';
+import { WaterHeaterManagement } from '@matter/main/clusters/water-heater-management';
+import { WaterHeaterMode } from '@matter/main/clusters/water-heater-mode';
+import { WaterHeaterManagementServer } from '@matter/main/behaviors/water-heater-management';
+import { WaterHeaterModeServer } from '@matter/main/behaviors/water-heater-mode';
 
 export class WaterHeater extends MatterbridgeEndpoint {
   /**
@@ -114,5 +119,42 @@ export class WaterHeater extends MatterbridgeEndpoint {
       currentMode: currentMode ?? 1,
     });
     return this;
+  }
+}
+
+export class MatterbridgeWaterHeaterManagementServer extends WaterHeaterManagementServer {
+  override boost({ boostInfo }: WaterHeaterManagement.BoostRequest): MaybePromise {
+    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
+    device.boost({ boostInfo });
+    device.log.info(`MatterbridgeWaterHeaterManagementServer boost called with: ${JSON.stringify(boostInfo)}`);
+    this.state.boostState = WaterHeaterManagement.BoostState.Active;
+    // The plugin is responsible for setting the device accordingly with the boostInfo of the boost command
+    // super.boost({ boostInfo });
+    // boost is not implemented in matter.js
+  }
+
+  override cancelBoost(): MaybePromise {
+    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
+    device.cancelBoost();
+    device.log.info(`MatterbridgeWaterHeaterManagementServer cancelBoost called`);
+    this.state.boostState = WaterHeaterManagement.BoostState.Inactive;
+    // The plugin is responsible for setting the device accordingly with the cancelBoost command
+    // super.cancelBoost();
+    // cancelBoost is not implemented in matter.js
+  }
+}
+
+export class MatterbridgeWaterHeaterModeServer extends WaterHeaterModeServer {
+  override changeToMode({ newMode }: ModeBase.ChangeToModeRequest): MaybePromise<ModeBase.ChangeToModeResponse> {
+    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
+    const supported = this.state.supportedModes.find((mode) => mode.mode === newMode);
+    if (!supported) {
+      device.log.error(`MatterbridgeWaterHeaterModeServer changeToMode called with unsupported newMode: ${newMode}`);
+      return { status: ModeBase.ModeChangeStatus.UnsupportedMode, statusText: 'Unsupported mode' };
+    }
+    device.changeToMode({ newMode });
+    this.state.currentMode = newMode;
+    device.log.info(`MatterbridgeWaterHeaterModeServer changeToMode called with newMode ${newMode} => ${supported.label}`);
+    return { status: ModeBase.ModeChangeStatus.Success, statusText: 'Success' };
   }
 }
